requirements.txt
# Core Dependencies
numpy==1.21.6
opencv-python==4.8.0.74
Pillow==9.5.0

# Deep Learning & Computer Vision
torch==1.13.0
torchvision==0.14.0
ultralytics==8.0.196  # YOLOv8 for person detection

# Audio Processing
pyaudio==0.2.13
pydub==0.25.1
webrtcvad==2.0.10  # Voice Activity Detection
librosa==0.10.0

# Speech Recognition (STT)
vosk==0.3.45  # Offline STT, very fast
SpeechRecognition==3.10.0

# Text-to-Speech (TTS)
pyttsx3==2.90  # Offline TTS, fast
gTTS==2.4.0  # Google TTS (backup option)

# LLM Integration
google-generativeai==0.3.1  # Gemini API
requests==2.31.0

# GUI for Face Display
pygame==2.5.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

